requirements.txt
boto3==1.9.82
botocore==1.12.82
Django==2.1.3
django-bootstrap-form==3.4
django-storages==1.7.1
docutils==0.14
jmespath==0.9.3
Pillow==5.4.1
python-dateutil==2.7.5
pytz==2018.9
s3transfer==0.1.13
six==1.12.0
urllib3==1.24.1
